requirements.txt
Flask~=2.0.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.28
psycopg2-binary==2.9.1
